More fixes, remove patient-invoices since invoices now deleted via visit.
